+++ src/app/Models/Attachable.php
@@ -1,0 +1,12 @@
+<?php
+namespace Afrittella\BackProject\Models;
+
+
+use Illuminate\Database\Eloquent\Model;
+
+class Attachable extends Model
+{
+    protected $guarded = [];
+
+
+}